[check] miss import

+++ Music/Plugins/audioandvideo.py
@@ -9,29 +9,37 @@ from pyrogram.types import (
     InputMediaAudio,
     Message,
 )
-from yt_dlp import YoutubeDL
-from pyrogram import Client, filters
-import wget
+
 import os
-from os import path
-from Music.MusicUtilities.helpers.thumbnails import down_thumb
-from Music import app
-from Music.MusicUtilities.helpers.inline import others_markup, play_markup
-from pykeyboard import InlineKeyboard
-from pyrogram import filters
-from pyrogram.types import InlineKeyboardButton
+import wget
 import yt_dlp
 import asyncio
+
+from os import path
 from PIL import Image
+from yt_dlp import YoutubeDL
 from hachoir.metadata import extractMetadata
 from hachoir.parser import createParser
+from pyrogram import Client, filters
+from pyrogram.types import InlineKeyboardButton
+
+from Music import app
+from ..MusicUtilities.helpers.thumbnails import down_thumb
+from Music.MusicUtilities.helpers.inline import play_markup, others_markup
+
+from pykeyboard import InlineKeyboard
+
+
 user_time = {}
 youtube_next_fetch = 0
 flex = {}
+
+
 @Client.on_callback_query(filters.regex(pattern=r"other"))
 async def closesmex(_,CallbackQuery):
     callback_data = CallbackQuery.data.strip()
     callback_request = callback_data.split(None, 1)[1]
+    await CallbackQuery.answer("menu opened")
     userid = CallbackQuery.from_user.id 
     videoid, user_id = callback_request.split("|")    
     buttons = others_markup(videoid, user_id)
@@ -42,10 +50,12 @@ async def closesmex(_,CallbackQuery):
 async def goback(_,CallbackQuery):
     callback_data = CallbackQuery.data.strip()
     callback_request = callback_data.split(None, 1)[1]
+    await CallbackQuery.answer("menu closed")
     userid = CallbackQuery.from_user.id 
     videoid, user_id = callback_request.split("|")    
     buttons = play_markup(videoid, user_id)
     await CallbackQuery.edit_message_reply_markup(reply_markup=InlineKeyboardMarkup(buttons))
+
 
 @Client.on_callback_query(filters.regex(pattern=r"good"))
 async def good(_,CallbackQuery):
@@ -54,17 +64,29 @@ async def good(_,CallbackQuery):
     userid = CallbackQuery.from_user.id 
     videoid, user_id = callback_request.split("|")    
     buttons = others_markup(videoid, user_id)
-    await CallbackQuery.edit_message_reply_markup(reply_markup=InlineKeyboardMarkup(buttons))    
-    
+    await CallbackQuery.edit_message_reply_markup(reply_markup=InlineKeyboardMarkup(buttons))
+
+
 @Client.on_callback_query(filters.regex("close"))
-async def closed(_, query: CallbackQuery):
+async def close_admin(_, query: CallbackQuery):
+    a = await _.get_chat_member(query.message.chat.id, query.from_user.id)
+    if not a.can_manage_voice_chats:
+        return await query.answer("💡 Only admin with manage video chat permission that can tap this button !", show_alert=True)
     await query.message.delete()
-    await query.answer()  
+    await query.answer()
+
+
+@Client.on_callback_query(filters.regex("cls"))
+async def close_user(_, query: CallbackQuery):
+    await query.message.delete()
+    await query.answer()
+
 
 @Client.on_callback_query(filters.regex(pattern=r"down"))
 async def down(_,CallbackQuery):
-    await CallbackQuery.answer()  
-    
+    await CallbackQuery.answer()
+
+
 @Client.on_callback_query(filters.regex(pattern=r"gets"))
 async def getspy(_,CallbackQuery):  
     callback_data = CallbackQuery.data.strip()
@@ -76,7 +98,7 @@ async def getspy(_,CallbackQuery):
     try:
         if userLastDownloadTime > datetime.now():
             wait_time = round((userLastDownloadTime - datetime.now()).total_seconds() / 60, 2)
-            return await CallbackQuery.message.reply_text(f"Okay {Name}, Fast AF Boiii.. `Wait for {wait_time} min(s) before next download request")
+            return await CallbackQuery.message.reply_text(f"hai {Name}, please wait for {wait_time} min(s) before do next download request.")
     except:
         pass
     url = (f"https://www.youtube.com/watch?v={videoid}")
@@ -87,7 +109,7 @@ async def getspy(_,CallbackQuery):
         user_time[userid] = now + \
                                      timedelta(minutes=youtube_next_fetch)
     except Exception:
-        return await CallbackQuery.message.reply_text("Failed To Fetch Data...")
+        return await CallbackQuery.message.reply_text("❌ failed to fetch data")
     j = 0
     a = 0
     b = 0
@@ -150,10 +172,10 @@ async def getspy(_,CallbackQuery):
                     f += 1
                     a6 = InlineKeyboardButton(text=f"(2160)p 📹 {humanbytes(x['filesize'])}", callback_data=f"ytdata video||{x['format_id']}||{videoid}")   
         else:
-            return await CallbackQuery.message.reply_text("Video Format Not Found.")
-    universal = InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+            return await CallbackQuery.message.reply_text("❌ video format not found.")
+    universal = InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
     if j == 0:
-        return await CallbackQuery.message.reply_text("Video Format Not Found..")
+        return await CallbackQuery.message.reply_text("❌ video format not found.")
     elif j == 1:
         key = InlineKeyboardMarkup(
             [
@@ -161,8 +183,8 @@ async def getspy(_,CallbackQuery):
                     a1,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )
@@ -174,8 +196,8 @@ async def getspy(_,CallbackQuery):
                     a2,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )  
@@ -190,8 +212,8 @@ async def getspy(_,CallbackQuery):
                     a3,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         ) 
@@ -207,8 +229,8 @@ async def getspy(_,CallbackQuery):
                     a4,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )    
@@ -227,8 +249,8 @@ async def getspy(_,CallbackQuery):
                     a5,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )  
@@ -248,13 +270,13 @@ async def getspy(_,CallbackQuery):
                     a6,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )    
     else:
-        return await CallbackQuery.message.reply_text("Video Format Not Found....")
+        return await CallbackQuery.message.reply_text("❌ video format not found.")
     await CallbackQuery.edit_message_reply_markup(reply_markup=key)
 
     
@@ -279,8 +301,8 @@ async def ytdata(_,CallbackQuery):
                     a2,
                 ],
                 [
-                    InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                    InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                    InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                    InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
                 ]    
             ]
         )
@@ -290,7 +312,7 @@ async def ytdata(_,CallbackQuery):
 inl = InlineKeyboardMarkup(
             [
                 [
-                    InlineKeyboardButton(text="Downloading......", callback_data=f'down')
+                    InlineKeyboardButton(text="📥 Downloading...", callback_data=f'down')
                 ]   
             ]
         )
@@ -298,7 +320,7 @@ inl = InlineKeyboardMarkup(
 upl = InlineKeyboardMarkup(
             [
                 [
-                    InlineKeyboardButton(text="Uploading......", callback_data=f'down')
+                    InlineKeyboardButton(text="📤 Uploading...", callback_data=f'down')
                 ]   
             ]
         )
@@ -306,17 +328,17 @@ upl = InlineKeyboardMarkup(
 def inl_mark(videoid, user_id):
     buttons= [
             [
-                InlineKeyboardButton(text="Download or Upload Failed......", callback_data=f'down')
+                InlineKeyboardButton(text="❌ download or upload failed.", callback_data=f'down')
             ],
             [
-                InlineKeyboardButton(text="⬅️ Go Back", callback_data=f'good {videoid}|{user_id}'),
-                InlineKeyboardButton(text="🗑 Close Menu", callback_data=f'close2')
+                InlineKeyboardButton(text="⬅️ Back", callback_data=f'good {videoid}|{user_id}'),
+                InlineKeyboardButton(text="🗑 Close", callback_data=f'close2')
             ],
         ]
     return buttons 
 
 
-ytdl_opts = {"format" : "bestaudio", "quiet":True}
+ytdl_opts = {"format" : "bestaudio/best", "quiet":True}
 
 
 @Client.on_callback_query(filters.regex(pattern=r"boom"))
@@ -339,27 +361,28 @@ async def boom(_,CallbackQuery):
     except Exception as e:
         buttons = inl_mark(videoid, user_id)
         await CallbackQuery.edit_message_reply_markup(reply_markup=InlineKeyboardMarkup(buttons)) 
-    mystic = await CallbackQuery.edit_message_text("Download Started\n\nDownloading speed could be slow. Please hold on..", reply_markup = inl)
+    mystic = await CallbackQuery.edit_message_text("📥 **download started** !\n\n» Downloading speed could be slow, please wait for a while...", reply_markup = inl)
     fetched = f"""
-**Track Downloaded**
-❇️**Title:** {x["title"]}
-⏳**Duration:** {round(x["duration"] / 60)} Mins
-👀**Views:** __{x["view_count"]}__
-🎥**Channel Name:** {x["uploader"]}
-__Youtube Inline Download Powered By Music__ """    
+💡 **Media Downloaded.**
+
+🏷 **Name:** {x["title"]}
+⏱ **Duration:** `{round(x["duration"] / 60)}` min(s)
+📣 **Channel:** {x["uploader"]}
+
+⚡️ __Powered by Steady Music AI__"""    
     link = (x["webpage_url"])
     channel = (x["channel_url"])
     perf = (x['uploader'])
     thumbnail_url = (x["thumbnail"])
-    filext = "%(title)s.png"
+    filext = "%(title)s.jpg"
     main =  os.getcwd()
     img = wget.download(thumbnail_url)
     im = Image.open(img).convert("RGB")
     output_directory = os.path.join(os.getcwd(), "search", str(userid))
     if not os.path.isdir(output_directory):
         os.makedirs(output_directory)
-    thumb_image_path = f"{output_directory}.png"
-    im.save(thumb_image_path,"png")
+    thumb_image_path = f"{output_directory}.jpg"
+    im.save(thumb_image_path,"jpeg")
     print(thumb_image_path)
     width = 0
     height = 0
@@ -378,7 +401,7 @@ __Youtube Inline Download Powered By Music__ """
             img.resize((320, height))
         else:
             img.resize((90, height))
-        img.save(thumb_image_path, "png")
+        img.save(thumb_image_path, "JPEG")
     audio_command = [
         "yt-dlp",
         "-c",
@@ -437,21 +460,18 @@ __Youtube Inline Download Powered By Music__ """
     if med:
         loop.create_task(send_file(CallbackQuery, med, filename, videoid, user_id, link, channel))
     else:
-        print("med not found")
+        print("media not found")
     
 def p_mark(link, channel):
     buttons= [
             [
-                InlineKeyboardButton(text="Watch on Youtube", url=f'{link}')
-            ],
-            [ 
-                InlineKeyboardButton(text="Visit Youtube Channel", url=f'{channel}')
+                InlineKeyboardButton(text="watch on youtube", url=f'{link}')
             ],
         ]
     return buttons    
     
 async def send_file(CallbackQuery, med, filename, videoid, user_id, link, channel):
-    await CallbackQuery.edit_message_text("Upload Started\n\nUploading speed could be slow. Please hold on..", reply_markup = upl)
+    await CallbackQuery.edit_message_text("📤 **upload started** !\n\n» Uploading speed could be slow, please wait for a while...", reply_markup = upl)
     try:
         await app.send_chat_action(chat_id=CallbackQuery.message.chat.id, action="upload_document")
         buttons = p_mark(link, channel)
@@ -508,7 +528,7 @@ def duration(vid_file_path):
             if 'duration' in s:
                 return float(s['duration'])
 
-    raise Exception('duration Not found')
+    raise Exception('duration not found')
     
     
 def humanbytes(num, suffix='B'):
